+++ skiworld/skiworld-ejb/src/main/java/Service/TrackEJBRemote.java
@@ -14,7 +14,7 @@ public interface TrackEJBRemote {
 	public void addTrack(Track t);
 	public void deleteTrack(Track t);
 	public void updateTrack(Track t);
-	public Track findTrackById(int idTrack);
+	public Track findTrackById(Long idTrack);
 	public List<Track> findAll();
 	public Track findTrackByDiff(String diff);
 }
